switch to folders for versions
