+++ packages/dd-trace/src/service-naming/schemas/util.js
@@ -1,0 +1,5 @@
+function identityService (service) {
+  return service
+}
+
+module.exports = { identityService }
