+++ packages/dd-trace/src/service-naming/schemas/v0/index.js
@@ -1,0 +1,4 @@
+const SchemaDefinition = require('../definition')
+const messaging = require('./messaging')
+
+module.exports = new SchemaDefinition({ messaging })
