--- packages/dd-trace/src/service-naming/schemas/v1.js
@@ -1,58 +1,0 @@
-const SchemaDefinition = require('./definition')
-
-function identityService (service) {
-  return service
-}
-
-const amqpInbound = {
-  opName: () => 'amqp.process',
-  serviceName: identityService
-}
-
-const amqpOutbound = {
-  opName: () => 'amqp.send',
-  serviceName: identityService
-}
-
-const schema = {
-  messaging: {
-    outbound: {
-      amqplib: amqpOutbound,
-      amqp10: amqpOutbound,
-      'google-cloud-pubsub': {
-        opName: () => 'gcp.pubsub.send',
-        serviceName: identityService
-      },
-      kafkajs: {
-        opName: () => 'kafka.send',
-        serviceName: identityService
-      },
-      rhea: amqpOutbound
-    },
-    inbound: {
-      amqplib: amqpInbound,
-      amqp10: amqpInbound,
-      'google-cloud-pubsub': {
-        opName: () => 'gcp.pubsub.process',
-        serviceName: identityService
-      },
-      kafkajs: {
-        opName: () => 'kafka.process',
-        serviceName: identityService
-      },
-      rhea: amqpInbound
-    },
-    controlPlane: {
-      amqplib: {
-        opName: () => 'amqp.command',
-        serviceName: identityService
-      },
-      'google-cloud-pubsub': {
-        opName: () => 'gcp.pubsub.request',
-        serviceName: identityService
-      }
-    }
-  }
-}
-
-module.exports = new SchemaDefinition(schema)
